+++ src/main/java/gg/projecteden/nexus/features/commands/staff/BurnCommand.java
@@ -1,11 +1,11 @@
 package gg.projecteden.nexus.features.commands.staff;
 
+import gg.projecteden.api.common.utils.TimeUtils.TickTime;
 import gg.projecteden.nexus.framework.commands.models.CustomCommand;
 import gg.projecteden.nexus.framework.commands.models.annotations.Path;
 import gg.projecteden.nexus.framework.commands.models.annotations.Permission;
 import gg.projecteden.nexus.framework.commands.models.annotations.Permission.Group;
 import gg.projecteden.nexus.framework.commands.models.events.CommandEvent;
-import gg.projecteden.utils.TimeUtils.TickTime;
 import org.bukkit.entity.Player;
 
 @Permission(Group.SENIOR_STAFF)
@@ -17,7 +17,7 @@ public class BurnCommand extends CustomCommand {
 
 	@Path("<player> <seconds>")
 	public void burn(Player player, int seconds) {
-		player.setFireTicks(TickTime.SECOND.x(seconds));
+		player.setFireTicks((int) TickTime.SECOND.x(seconds));
 		send(PREFIX + "&3Set &e" + player.getName() + "&3 on fire for &e" + seconds + plural(" &3second", seconds));
 	}
 }
