+++ src/main/java/gg/projecteden/nexus/features/minigames/lobby/ActionBar.java
@@ -5,14 +5,14 @@ import gg.projecteden.nexus.features.minigames.commands.MinigamesCommand.Minigam
 import gg.projecteden.nexus.features.minigames.utils.MinigameNight.NextMGN;
 import gg.projecteden.nexus.utils.ActionBarUtils;
 import gg.projecteden.nexus.utils.Tasks;
-import gg.projecteden.utils.TimeUtils.TickTime;
+import gg.projecteden.api.common.utils.TimeUtils.TickTime;
 import net.citizensnpcs.api.npc.NPC;
 import net.md_5.bungee.api.ChatColor;
 import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -20,7 +20,7 @@ import static gg.projecteden.nexus.utils.CitizensUtils.isNPC;
 import static gg.projecteden.nexus.utils.StringUtils.camelCase;
 
 public class ActionBar {
-	private static final int DELAY = TickTime.SECOND.x(5);
+	private static final long DELAY = TickTime.SECOND.x(5);
 
 	private List<String> messages = new ArrayList<>() {{
 		add("&3You are currently in the &eMinigame Lobby&3!");
@@ -37,7 +37,7 @@ public class ActionBar {
 
 	public ActionBar() {
 		Tasks.repeat(5, messages.size() * DELAY, () -> {
-			AtomicInteger wait = new AtomicInteger(0);
+			AtomicLong wait = new AtomicLong(0);
 			messages.iterator().forEachRemaining(message ->
 					Tasks.wait(wait.getAndAdd(DELAY), () ->
 							Minigames.getWorld().getPlayers().forEach(player -> {
