+++ src/main/java/gg/projecteden/nexus/features/minigames/models/matchdata/UHCMatchData.java
@@ -5,7 +5,7 @@ import gg.projecteden.nexus.features.minigames.models.Match;
 import gg.projecteden.nexus.features.minigames.models.MatchData;
 import gg.projecteden.nexus.features.minigames.models.Minigamer;
 import gg.projecteden.nexus.features.minigames.models.annotations.MatchDataFor;
-import gg.projecteden.utils.TimeUtils.Timespan;
+import gg.projecteden.api.common.utils.TimeUtils.Timespan;
 import lombok.Data;
 
 import java.time.LocalDateTime;
@@ -17,7 +17,7 @@ import java.util.UUID;
 @MatchDataFor(UHC.class)
 public class UHCMatchData extends MatchData {
 	private LocalDateTime startTime = LocalDateTime.now(); // placeholder, is updated on true game start
-	private final Map<UUID, Integer> timeAlive = new HashMap<>();
+	private final Map<UUID, Long> timeAlive = new HashMap<>();
 
 	public UHCMatchData(Match match) {
 		super(match);
