+++ src/main/java/gg/projecteden/nexus/features/particles/effects/DiscoEffect.java
@@ -6,7 +6,7 @@ import gg.projecteden.nexus.framework.exceptions.postconfigured.InvalidInputExce
 import gg.projecteden.nexus.models.particle.ParticleService;
 import gg.projecteden.nexus.utils.RandomUtils;
 import gg.projecteden.nexus.utils.Tasks;
-import gg.projecteden.utils.TimeUtils.TickTime;
+import gg.projecteden.api.common.utils.TimeUtils.TickTime;
 import lombok.Builder;
 import lombok.Getter;
 import org.bukkit.Color;
@@ -23,7 +23,7 @@ public class DiscoEffect {
 
 	@Builder(buildMethodName = "start")
 	public DiscoEffect(HumanEntity player, Location location, boolean updateLoc, Vector updateVector,
-					   Direction direction, int ticks, boolean rainbow, RainbowOption rainbowOption,
+					   Direction direction, long ticks, boolean rainbow, RainbowOption rainbowOption,
 					   Particle lineParticle, int maxLines, int lineDensity, int lineLength, double lineSpeed, boolean lineRainbow, Color lineColor, double disX1, double disY1, double disZ1,
 					   Particle sphereParticle, double sphereRadius, int sphereDensity, double sphereSpeed, boolean sphereRainbow, Color sphereColor, double disX2, double disY2, double disZ2,
 					   int startDelay, int pulseDelay) {
@@ -112,7 +112,7 @@ public class DiscoEffect {
 		Location finalLocation = location;
 		boolean finalUpdateLoc = updateLoc;
 		Vector finalUpdateVector = updateVector;
-		int finalTicks = ticks;
+		long finalTicks = ticks;
 		AtomicInteger ticksElapsed = new AtomicInteger(0);
 
 		taskId = Tasks.repeat(startDelay, pulseDelay, () -> {
