+++ src/main/java/gg/projecteden/nexus/features/particles/effects/NyanCatEffect.java
@@ -4,7 +4,7 @@ import gg.projecteden.nexus.features.particles.ParticleUtils;
 import gg.projecteden.nexus.framework.exceptions.postconfigured.InvalidInputException;
 import gg.projecteden.nexus.models.particle.ParticleService;
 import gg.projecteden.nexus.utils.Tasks;
-import gg.projecteden.utils.TimeUtils.TickTime;
+import gg.projecteden.api.common.utils.TimeUtils.TickTime;
 import lombok.Builder;
 import lombok.Getter;
 import org.bukkit.Location;
@@ -19,14 +19,14 @@ public class NyanCatEffect {
 	private int taskId;
 
 	@Builder(buildMethodName = "start")
-	public NyanCatEffect(HumanEntity player, int ticks, int startDelay, int pulseDelay) {
+	public NyanCatEffect(HumanEntity player, long ticks, int startDelay, int pulseDelay) {
 
 		if (player == null) throw new InvalidInputException("No player was provided");
 
 		if (pulseDelay < 1) pulseDelay = 1;
 		if (ticks == 0) ticks = TickTime.SECOND.x(5);
 
-		int finalTicks = ticks;
+		long finalTicks = ticks;
 		AtomicInteger ticksElapsed = new AtomicInteger(0);
 
 		taskId = Tasks.repeat(startDelay, pulseDelay, () -> {
