+++ src/main/java/gg/projecteden/nexus/features/particles/effects/SphereEffect.java
@@ -5,7 +5,7 @@ import gg.projecteden.nexus.features.particles.ParticleUtils;
 import gg.projecteden.nexus.framework.exceptions.postconfigured.InvalidInputException;
 import gg.projecteden.nexus.models.particle.ParticleService;
 import gg.projecteden.nexus.utils.Tasks;
-import gg.projecteden.utils.TimeUtils.TickTime;
+import gg.projecteden.api.common.utils.TimeUtils.TickTime;
 import lombok.Builder;
 import lombok.Getter;
 import org.bukkit.Color;
@@ -23,7 +23,7 @@ public class SphereEffect {
 
 	@Builder(buildMethodName = "start")
 	public SphereEffect(HumanEntity player, Location location, boolean updateLoc, Vector updateVector, Particle particle,
-						boolean rainbow, Color color, int count, int density, double radius, int ticks, double speed,
+						boolean rainbow, Color color, int count, int density, double radius, long ticks, double speed,
 						double disX, double disY, double disZ, int startDelay, int pulseDelay) {
 
 		if (player != null && location == null)
@@ -60,7 +60,7 @@ public class SphereEffect {
 		double finalRadius = radius;
 		int finalCount = count;
 		double finalSpeed = speed;
-		int finalTicks = ticks;
+		long finalTicks = ticks;
 		Particle finalParticle = particle;
 		Location finalLocation = location;
 		boolean finalUpdateLoc = updateLoc;
@@ -91,7 +91,6 @@ public class SphereEffect {
 			if (finalUpdateLoc)
 				newLoc = player.getLocation().add(finalUpdateVector);
 			newLoc.add(0, 1, 0);
-
 
 			t.updateAndGet(v -> v + Math.PI / (double) finalDensity);
