+++ src/main/java/gg/projecteden/nexus/features/particles/effects/StormEffect.java
@@ -5,7 +5,7 @@ import gg.projecteden.nexus.framework.exceptions.postconfigured.InvalidInputExce
 import gg.projecteden.nexus.models.particle.ParticleService;
 import gg.projecteden.nexus.utils.RandomUtils;
 import gg.projecteden.nexus.utils.Tasks;
-import gg.projecteden.utils.TimeUtils.TickTime;
+import gg.projecteden.api.common.utils.TimeUtils.TickTime;
 import lombok.Builder;
 import lombok.Getter;
 import org.bukkit.Color;
@@ -22,7 +22,7 @@ public class StormEffect {
 
 	@Builder(buildMethodName = "start")
 	public StormEffect(HumanEntity player, Location location, boolean updateLoc, Vector updateVector, RainPartile rainParticle,
-					   int density, double radius, int ticks, int startDelay, int pulseDelay) {
+					   int density, double radius, long ticks, int startDelay, int pulseDelay) {
 
 		Particle cloudParticle = Particle.REDSTONE;
 		Color cloudColor = Color.fromRGB(127, 127, 127);
@@ -46,7 +46,7 @@ public class StormEffect {
 		double rainRadius = radius - 0.25;
 		double finalCloudRadius = radius;
 		int finalDensity = density;
-		int finalTicks = ticks;
+		long finalTicks = ticks;
 		RainPartile finalRainParticle = rainParticle;
 		Location finalLocation = location;
 		boolean finalUpdateLoc = updateLoc;
